LayoutTests/inspector/debugger/source-frame-expected.txt

CONSOLE MESSAGE: line 10: test error message
Tests that it's possible to set breakpoint in source frame, and that source frame displays breakpoints and console errors.

Debugger was enabled.

Running: testSetBreakpoint
Script source was shown.
Breakpoint added to source frame at line 14

Running: testConsoleMessage
Script source was shown.
Message added to source frame: test error message
Messages cleared in source frame.

Running: testShowResource
Debugger was disabled.
